Page through all checkins for current day
Previously only the first page was being used and the date was not being
checked, so the last 20 results were always being displayed. Now results
are only displayed for the current day and the next page of results will
be called if the last item on the current page is the current date.