Use application cache to store checkins
Now since we have to make several requests to retrieve all checkins for
the day it is better to cache all the checkins. That way we can compare
the first barcode in the cache with the first barcode from the first
page of results. if they match then we don't need to make any further
requests.