Use setTimeout to refresh results
When using setInterval then the time taken to retrieve the results is not
included in the time. With setTimeout we can set the time after the
results have be retrieved so if retrieving the results takes a long time
then when we display the results we won't also display that we have
already waited a length of time and are getting ready to retrieve the
results again.