Create store through webhook and change data structure
The request for the checkin count no longer retrieves the data from the
city. Instead it only reads the current values in the atom. The
structure of the atom has also been changed. It is now a map where the
barcodes are the keys and the value is the group id. This way when we
retrieve new checkins from the city we have something easy to compare
against to determine if we have already recorded the checkin in the
atom.